[Update] 抽象BasePermission (#2710)
* [Update] AssetPermission/RemoteAppPermission抽象BasePermission

* [Update] Perms模块添加迁移文件

* [Update] Perms删除多余迁移文件

* [Update] Perms重新生成RemoteAppPermission迁移文件
